next_permutation leetcode
This is a solution to next_permutation problem on leetcode with python. It traverses the list from reverse and goes on until it finds an element which is not in descending order and assigns a pivot position to it. The pivot element is then swapped with another element which is bigger than it. The list is then sorted from the pivot element to get the next permutation.